+ Thêm chức năng xem danh sách học sinh + Thêm chức năng thêm học sinh + Thêm chức năng tìm kiếm học sinh + Thêm chức năng lọc học sinh + Thêm chức năng xem thông tin chi tiết học sinh + Thêm chức năng chỉnh sửa thông tin học sinh + Thêm chức năng xem danh sách thông báo + Thêm chức năng xem chi tiết thông báo + Thêm chức năng thêm thông báo + Thêm chức năng chỉnh sửa thông báo + Thêm chức năng tìm kiếm thông báo

+++ app/src/main/java/tkpm/doan/student/data/models/Notification.java
@@ -12,7 +12,8 @@ public class Notification {
 
     // Getter Methods
 
-    public float getNotificationId() {
+
+    public int getNotificationId() {
         return NotificationId;
     }
